Restore backward compatibility
