+++ common/kotlinx-coroutines-core-common/src/Dispatchers.common.kt
@@ -2,10 +2,6 @@
  * Copyright 2016-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-/*
- * Copyright 2016-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
- */
-
 package kotlinx.coroutines.experimental
 
 import kotlinx.coroutines.experimental.internal.*
