+++ core/kotlinx-coroutines-core/src/Dispatchers.kt
@@ -2,6 +2,8 @@
  * Copyright 2016-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
+@file:Suppress("unused")
+
 package kotlinx.coroutines.experimental
 
 import kotlinx.coroutines.experimental.scheduling.*
